refactor: move journal-entry builder from src/domain to tests/fixtures
